CMS-OSD separation
