+++ src/main/io/canvas.h
@@ -1,0 +1,3 @@
+#pragma once
+
+screenFnVTable_t *canvasInit(void);
