+++ src/main/io/ledstrip.c
@@ -62,8 +62,8 @@
 #include "io/gimbal.h"
 #include "io/serial.h"
 #include "io/gps.h"
-#include "io/osd.h"
-#include "io/vtx.h"
+//#include "io/osd.h"
+//#include "io/vtx.h"
 
 #include "flight/failsafe.h"
 #include "flight/mixer.h"
