+++ src/main/telemetry/smartport.c
@@ -36,9 +36,9 @@
 #include "io/gps.h"
 #include "io/gimbal.h"
 #include "io/serial.h"
-#include "io/ledstrip.h"
-#include "io/osd.h"
-#include "io/vtx.h"
+//#include "io/ledstrip.h"
+//#include "io/osd.h"
+//#include "io/vtx.h"
 
 #include "sensors/boardalignment.h"
 #include "sensors/sensors.h"
